IU_TimControls and Mockup3
Add hidden controls management and sample.
When a form gives focus to a Tim Control it must check if the control is visible or not. If the control is not visible it must call moveNext or movePrevious for selecting previous or next control. The best way is to call the KeyPress event manager of the form again (see the sample gives with the setFocus method of IU_TimControls unit